// framework/ui/common/src/lib/domain-controls/model/input-select-items.interface.ts
/*
 * Copyright(c) RIB Software GmbH
 */

/**
 * This object stores input select items information.
 */
export interface IInputSelectItems {
	/**
	 * Description.
	 */
	description: string;

	/**
	 * Id.
	 */
	id: number;

	/**
	 * IsLive.
	 */
	isLive: boolean;

	/**
	 * Remark.
	 */
	remark: string;

	/**
	 * Sorting.
	 */
	sorting: number;

	/**
	 * Version.
	 */
	version: number;
}